+++ config/jsoref-spellchecker/whitelist.words
@@ -120,7 +120,6 @@ blockandinlinetags
 blockcomment
 blockcommentposition
 blockcommentstyle
-blockindentation
 blockquote
 blog
 blogspot
@@ -589,7 +588,6 @@ importorder
 importrule
 imprt
 Incompatabilities
-indentcontinuationlines
 Inet
 infinispan
 initializedfields
