+++ src/xdocs/google_style.xml
@@ -161,7 +161,7 @@
                 </td>
                 <td>
                   <a href="styleguides/google-java-style-20220203/javaguide.html#s2.1-file-name">
-                    2.1 File Name</a>
+                    2.1 File name</a>
                 </td>
                 <td>
                   <span class="wrapper inline">
@@ -211,7 +211,7 @@
                 </td>
                 <td>
                   <a href="styleguides/google-java-style-20220203/javaguide.html#s2.3.1-whitespace-characters">
-                    2.3.1 Whitespace Characters</a>
+                    2.3.1 Whitespace characters</a>
                 </td>
                 <td>
                   <span class="wrapper inline">
@@ -335,7 +335,7 @@
                 </td>
                 <td>
                   <a href="styleguides/google-java-style-20220203/javaguide.html#s3.2-package-statement">
-                    3.2 Package Statement</a>
+                    3.2 Package statement</a>
                 </td>
                 <td>
                   <span class="wrapper inline">
@@ -578,13 +578,10 @@
                   <a href="https://github.com/search?q=path%3Asrc%2Fmain%2Fresources%20path%3A**%2Fgoogle_checks.xml+repo%3Acheckstyle%2Fcheckstyle+OverloadMethodsDeclarationOrder">
                     config</a>
                   <br />
-                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter3filestructure/rule3421overloadsplit/OverloadMethodsDeclarationOrderTest.java">
-                    test</a>
-                  <br />
                   <a href="https://github.com/search?q=path%3Asrc%2Fmain%2Fresources%20path%3A**%2Fgoogle_checks.xml+repo%3Acheckstyle%2Fcheckstyle+ConstructorsDeclarationGrouping">
                     config</a>
                   <br />
-                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter3filestructure/rule3421overloadsplit/ConstructorsDeclarationGroupingTest.java">
+                  <a href="https://github.com/checkstyle/checkstyle/blob/master/src/it/java/com/google/checkstyle/test/chapter3filestructure/rule3421overloadsplit/OverloadsNeverSplitTest.java">
                     test</a>
                   <br />
                 </td>
